Break the circular dependency by extending a pluggable map
