+++ examples/extent-constrained.js
@@ -2,7 +2,7 @@ import Map from '../src/ol/Map.js';
 import View from '../src/ol/View.js';
 import TileLayer from '../src/ol/layer/Tile.js';
 import OSM from '../src/ol/source/OSM.js';
-import {defaults as defaultControls} from '../src/ol/control/util';
+import {defaults as defaultControls} from '../src/ol/control.js';
 import ZoomSlider from '../src/ol/control/ZoomSlider';
 
 const view = new View({
